Changed sizeInBits field to int.  This seems more suitable (int should be able to hold the total bit length of a parameter!) and prevents us from casting back to int all the time.

+++ hummingbird/commons/src/main/java/com/logica/hummingbird/spacesystemmodel/parameters/behaviours/AbstractFloatBehaviour.java
@@ -6,7 +6,7 @@ import com.logica.hummingbird.spacesystemmodel.exceptions.BitSetOperationExcepti
 
 public abstract class AbstractFloatBehaviour extends AbstractNumberBehaviour {
 
-	public AbstractFloatBehaviour(long sizeInBits) {
+	public AbstractFloatBehaviour(int sizeInBits) {
 		super(sizeInBits);
 	}
