+++ hummingbird/commons/src/main/java/com/logica/hummingbird/spacesystemmodel/parameters/behaviours/IntegerUnsignedBehaviour.java
@@ -11,7 +11,7 @@ import com.logica.hummingbird.util.BitSetUtility;
 public class IntegerUnsignedBehaviour extends AbstractIntegerBehaviour {
 	private static final Logger LOG = LoggerFactory.getLogger(IntegerUnsignedBehaviour.class);
 	
-	public IntegerUnsignedBehaviour(long sizeInBits) throws InvalidParameterTypeException {
+	public IntegerUnsignedBehaviour(int sizeInBits) throws InvalidParameterTypeException {
 		super(sizeInBits);
 		if(sizeInBits > 32) {
 			throw new InvalidParameterTypeException("Integer unsigned cannot be greater than 32-bits in size.");
