+++ hummingbird/commons/src/main/java/com/logica/hummingbird/spacesystemmodel/parameters/behaviours/ParameterBehaviour.java
@@ -6,7 +6,7 @@ import com.logica.hummingbird.spacesystemmodel.exceptions.BitSetOperationExcepti
 
 public interface ParameterBehaviour {
 	
-	public long getSizeIntBits();
+	public int getSizeIntBits();
 	
 	/**
 	 * Returns the raw binary data of the parameter as a BitSet
